Reset the database to initial state before each test
Many unit tests modify data within our firebase emulator database. These changes made during the test remain persistent, and tests are not indendent.

For example, consider a test in which a user joins a given ride. When the test is run again, the user already exists in that ride from the previous run, and thus the test fails.

This commit resets the database to prevent such issues. Before each test is executed, the database is set to an initial state containing the original data. As a result, tests remain independent between runs.

Based on https://stackoverflow.com/questions/68317339/testing-svelte-components-with-import-meta-env.